src/ghosts.shadows/api/requirements.txt
Flask
gradio
imaginairy
langchain
langchain_community
numpy
Requests
torch
transformers
docx2txt
pypdf
chromadb
tqdm
pipdeptree
watchdog
python-jose
passlib
aiosqlite
